Refactored plugin to it's own file
